requirements.txt
Flask>=2.0
SQLAlchemy>=2.0
psycopg2-binary>=2.8
python-dotenv>=1.0
Flask-SQLAlchemy>=3.0
Flask-CORS>=4.0
gunicorn